Moved errors to types crate, fixed docker make down

+++ Makefile
@@ -1,7 +1,7 @@
 up:
 		docker compose -f docker/docker-compose.yaml up
 down:
-		rm -r ./api/tmp/*
+		[ ! -e file ] || rm -r ./api/tmp/*
 		docker compose -f docker/docker-compose.yaml down
 build:
 		mkdir -p ./api/tmp
